Restructure project organization: - Create helpers folder - Convert DateRangeGenerator and IngredientsParser classes to helper functions and move them to the helpers folder - Rename helpers.py file and add it to the helpers folder - Adjust the rest of the code to accommodate these changes.
