+++ src/helpers/login_required_decorator.py
@@ -1,8 +1,4 @@
-import os
-# import requests
-import urllib.parse
-
-from flask import redirect, render_template, request, session
+from flask import redirect, session
 from functools import wraps
 
 def login_required(f):
