--- src/services/shopping_list_module/__init__.py
@@ -1,3 +1,0 @@
-from .shopping_list_service import ShoppingListService
-from .date_range_generator import DateRangeGenerator
-from .ingredient_parser import IngredientParser
